requirements.txt
google_api_python_client==2.125.0
google_auth_oauthlib==1.2.0
icalendar==5.0.11
protobuf==5.26.1
pytz==2023.3.post1
Requests==2.31.0
